add option to auto-launch browser on server
Add command-line argument --browser (shortened to -b) which will
auto-launch a web-browser pointing to the Klaus server via the Python
webbrowser module.

Note that we do things in the wrong order in that we launch the browser
and *then* start the server. In the absence of a cleaner way of getting
a function to run after server setup this approach actually works in
practice. If this doesn't Just Work for someone we can think again about
more complex options.